+++ src/js/main.js
@@ -1,4 +1,5 @@
-import renderControlElements from './modules/_controls';
+// import renderTripPoints from './modules/points';
+import renderControlElements from './modules/controls';
 
 renderControlElements(`filter`);
 renderControlElements(`sorting`);
